Refactor App code into TrainingCosts component
