+++ src/utilities.js
@@ -6,3 +6,15 @@ export const getNamesAndValues = (obj) => {
 
   return namesAndValues;
 };
+
+export const areInputsValid = (getValidationErrors, setErrors) => {
+  const validationErrors = getValidationErrors();
+
+  if (Object.keys(validationErrors).length > 0) {
+    setErrors(validationErrors);
+    return false;
+  } else {
+    setErrors({});
+    return true;
+  }
+};
